Move matmul dims check test
